+++ Components/SingleItem.jsx
@@ -26,6 +26,8 @@ const SingleItem = () => {
   };
 
   const [thumbsSwiper, setThumbsSwiper] = useState(null);
+
+
   const [cartItems, setCartItems] = useState([]);
   useEffect(() => {
     // Initialize cart items from local storage on component mount
@@ -35,17 +37,31 @@ const SingleItem = () => {
     }
   }, []);
 
-  const addToCart = (id, title, price, qtyValue) => {
-    const newCartItem = { id, title, price, qtyValue };
-    const updatedCartItems = [...cartItems, newCartItem];
+  const addToCart = (id, title, price, isDiscounted, discounted_price, discounted_percent, qtyValue, img_src, stock) => {
+    const existingCartItem2 = cartItems.find(item => item.id === id);
 
-    // Update state and local storage with new cart items
-    setCartItems(updatedCartItems);
-    localStorage.setItem("cartItems", JSON.stringify(updatedCartItems));
-    alert("Item added to cart successfully");
-    // window.location.reload();
-    // Update cart items count in local storage
-    // localStorage.setItem("cartItemsCount", updatedCartItems.length);
+    if (existingCartItem2) {
+      // Item with the same ID already exists in the cart
+      alert(`${title} is already in the cart.`);
+    } else {
+      // Item with the same ID doesn't exist, add it to the cart
+      const newCartItem = {
+        id,
+        title,
+        price,
+        isDiscounted,
+        discounted_price,
+        discounted_percent,
+        qtyValue,
+        img_src,
+        stock
+      };
+      const updatedCartItems = [...cartItems, newCartItem];
+
+      setCartItems(updatedCartItems);
+      localStorage.setItem("cartItems", JSON.stringify(updatedCartItems));
+      alert(`${title} added to cart successfully`);
+    }
   };
   return (
     <div className="  flex flex-col sm:flex-row mx-3 mt-5 sm:mx-5">
@@ -108,20 +124,30 @@ const SingleItem = () => {
         </div>}
         <div className="flex items-center gap-x-4 mt-3 mx-1 text-lg">
           <div className="flex justify-center items-center gap-x-2">
-            {item.isDiscounted ? <><span className="text-xl">{item.discounted_price}₹</span>
-              <span className="text-xs line-through">{item.price}₹</span></> : <span className="text-xl">{item.price}₹</span>}
+            {item.isDiscounted ? <><span className="text-xl">₹{item.discounted_price}</span>
+              <span className="text-xs line-through">₹{item.price}</span></> : <span className="text-xl">₹{item.price}</span>}
 
           </div>
-          <span className="hover:text-[#131b2e] hover:bg-white border whitespace-nowrap hover:border-[#232f3e] bg-[#232f3e] text-white cursor-pointer  font-bold py-2 px-4 rounded-full ">
+          {item.stock !== 0 ? <span className="hover:text-[#131b2e] hover:bg-white border whitespace-nowrap hover:border-[#232f3e] bg-[#232f3e] text-white cursor-pointer  font-bold py-2 px-4 rounded-full ">
             <button
               onClick={() => {
-                // You can adjust the item ID, title, and price as needed
-                addToCart(cartItems.length + 1, "Item Name", 1000, 1);
+                addToCart(
+                  item._id,
+                  item.title,
+                  item.price,
+                  item.isDiscounted,
+                  item.discounted_price,
+                  item.discounted_percent,
+                  1,
+                  item.main_img,
+                  item.stock
+                );
               }}
             >
               Add to cart
             </button>
-          </span>
+          </span> : <span className="text-sm text-[crimson]">This Item is out of Stock</span>}
+
         </div>
       </div>
     </div>
